Исправление кода согласно замечаниям ревьюера
