+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -1,10 +1,10 @@
 package ru.yandex.practicum.filmorate.model;
 
 import lombok.Data;
-import ru.yandex.practicum.filmorate.customvalidators.NoWhitespaceValidation;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Past;
+import javax.validation.constraints.Pattern;
 import java.time.LocalDate;
 
 @Data
@@ -14,7 +14,7 @@ public class User {
     @Email(message = "Неверно указан формат электронной почты")
     String email;
     @NotBlank(message = "логин пользователя не должен быть пустым")
-    @NoWhitespaceValidation(message = "в логине не должно содержаться пробелов")
+    @Pattern(regexp = "\\S*", message = "в логине не должно содержаться пробелов")
     String login;
     String name;
     @Past(message = "Дата дня рождения не может быть в будущем или сегодняшним днём")
