+++ src/test/java/ru/yandex/practicum/filmorate/GetUsersTest.java
@@ -1,0 +1,49 @@
+package ru.yandex.practicum.filmorate;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Order;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.RequestBuilder;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static ru.yandex.practicum.filmorate.Utils.readFile;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_CLASS)
+public class GetUsersTest {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @BeforeEach
+    public void beforeEach() throws Exception {
+        //Добавление нового пользователя
+        String body = readFile("src/test/resources/users/addUser.json");
+        RequestBuilder builder = MockMvcRequestBuilders
+                .post("/users")
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(body);
+        this.mockMvc.perform(builder)
+                .andExpect(status().isOk())
+                .andExpect(content().json(body));
+    }
+
+    @Test
+    @DisplayName("Получение списка всех пользователей")
+    public void getUsers() throws Exception {
+        RequestBuilder builder = MockMvcRequestBuilders
+                .get("/users");
+        this.mockMvc.perform(builder)
+                .andExpect(status().isOk());
+    }
+}
