+++ src/test/java/ru/yandex/practicum/filmorate/UpdateUserInfoWithIncorrectIdTest.java
@@ -1,0 +1,39 @@
+package ru.yandex.practicum.filmorate;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Order;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.RequestBuilder;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+
+import static org.hamcrest.Matchers.is;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static ru.yandex.practicum.filmorate.Utils.readFile;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+public class UpdateUserInfoWithIncorrectIdTest {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Test
+    @Order(13)
+    @DisplayName("Обновление информации по пользователю с несуществующим id")
+    public void updateUserInfoWithIncorrectId() throws Exception {
+        String body = readFile("src/test/resources/users/updateUserWithIncorrectId.json");
+        RequestBuilder builder = MockMvcRequestBuilders
+                .put("/users")
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(body);
+        this.mockMvc.perform(builder)
+                .andExpect(status().is(404))
+                .andExpect(jsonPath("$.message", is("Попытка обновить информацию по пользователю с несуществующим id")));
+    }
+}
